Move properties light dat iso`and àlignment wavelenth` to RayDataBuilder.
