+++ opossum_core/examples/workshop_09_phelix.rs
@@ -12,7 +12,7 @@ fn main() -> OpmResult<()> {
     let mut scenery = NodeGroup::new("PHELIX MainAmp");
 
     let light_data_builder =
-        LightDataBuilder::Geometric(RayDataBuilder::Collimated(CollimatedSrc::new(
+        LightDataBuilder::Geometric(RayDataSource::Collimated(CollimatedSrc::new(
             Grid::new(
                 Point2::new(Length::zero(), millimeter!(60.0)),
                 Point2::new(1, 5),
