+++ opossum_core/examples/workshop_11_ghostfocus.rs
@@ -8,7 +8,7 @@ use std::path::Path;
 fn main() -> OpmResult<()> {
     let mut scenery = NodeGroup::new("Ghostfocus demo");
     let light_data_builder =
-        LightDataBuilder::Geometric(RayDataBuilder::Collimated(CollimatedSrc::new(
+        LightDataBuilder::Geometric(RayDataSource::Collimated(CollimatedSrc::new(
             HexagonalTiling::new(millimeter!(15.0), 25, millimeter!(0.0, 0.))?.into(),
             General2DGaussian::new(
                 joule!(2.),
