+++ opossum_core/src/lightdata/light_data_builder.rs
@@ -7,11 +7,11 @@ use serde::{Deserialize, Serialize};
 use std::fmt::Display;
 use strum::EnumIter;
 
-use super::{LightData, energy_data_builder::EnergyDataBuilder, ray_data_builder::RayDataBuilder};
+use super::{LightData, energy_data_builder::EnergyDataBuilder, ray_data_source::RayDataSource};
 use crate::{
     energy_distributions::EnergyDistType,
     error::OpmResult,
-    lightdata::ray_data_builder::{CollimatedSrc, ImageSrc, PointSrc},
+    lightdata::ray_data_source::{CollimatedSrc, ImageSrc, PointSrc},
     position_distributions::PosDistType,
     spectral_distribution::SpecDistType,
     utils::default_from_name::DefaultFromName,
@@ -23,7 +23,7 @@ pub enum LightDataBuilder {
     /// Builder for the generation of [`LightData::Energy`].
     Energy(EnergyDataBuilder),
     /// Builder for the generation of [`LightData::Geometric`].
-    Geometric(RayDataBuilder),
+    Geometric(RayDataSource),
     // /// Dummy Fourier
     // Fourier,
 }
@@ -32,7 +32,7 @@ impl DefaultFromName for LightDataBuilder {}
 
 impl Default for LightDataBuilder {
     fn default() -> Self {
-        Self::Geometric(RayDataBuilder::default())
+        Self::Geometric(RayDataSource::default())
     }
 }
 
@@ -123,19 +123,19 @@ impl From<EnergyDataBuilder> for LightDataBuilder {
 }
 impl From<ImageSrc> for LightDataBuilder {
     fn from(value: ImageSrc) -> Self {
-        Self::Geometric(RayDataBuilder::Image(value))
+        Self::Geometric(RayDataSource::Image(value))
     }
 }
 
 impl From<PointSrc> for LightDataBuilder {
     fn from(value: PointSrc) -> Self {
-        Self::Geometric(RayDataBuilder::PointSrc(value))
+        Self::Geometric(RayDataSource::PointSrc(value))
     }
 }
 
 impl From<CollimatedSrc> for LightDataBuilder {
     fn from(value: CollimatedSrc) -> Self {
-        Self::Geometric(RayDataBuilder::Collimated(value))
+        Self::Geometric(RayDataSource::Collimated(value))
     }
 }
 
@@ -232,7 +232,7 @@ mod tests {
         ));
         let light_data = light_data_builder.build().unwrap();
         assert!(matches!(light_data, LightData::Energy(_)));
-        let light_data_builder = LightDataBuilder::Geometric(RayDataBuilder::Raw(Rays::default()));
+        let light_data_builder = LightDataBuilder::Geometric(RayDataSource::Raw(Rays::default()));
         let light_data = light_data_builder.build().unwrap();
         assert!(matches!(light_data, LightData::Geometric(_)));
     }
