+++ opossum_core/src/nodes/source_helper.rs
@@ -7,7 +7,7 @@ use crate::{
     error::{OpmResult, OpossumError},
     lightdata::{
         light_data_builder::LightDataBuilder,
-        ray_data_builder::{CollimatedSrc, PointSrc, RayDataBuilder},
+        ray_data_source::{CollimatedSrc, PointSrc, RayDataSource},
     },
     millimeter, nanometer,
     position_distributions::{Grid, Hexapolar},
@@ -33,7 +33,7 @@ pub fn round_collimated_ray_source(
     nr_of_rings: u8,
 ) -> OpmResult<Source> {
     let light_data_builder =
-        LightDataBuilder::Geometric(RayDataBuilder::Collimated(CollimatedSrc::new(
+        LightDataBuilder::Geometric(RayDataSource::Collimated(CollimatedSrc::new(
             Hexapolar::new(radius, nr_of_rings)?.into(),
             UniformDist::new(energy)?.into(),
             LaserLines::new(vec![(nanometer!(1000.0), 1.0)])?.into(),
@@ -58,7 +58,7 @@ pub fn collimated_line_ray_source(
     nr_of_points_y: usize,
 ) -> OpmResult<Source> {
     let light_data_builder =
-        LightDataBuilder::Geometric(RayDataBuilder::Collimated(CollimatedSrc::new(
+        LightDataBuilder::Geometric(RayDataSource::Collimated(CollimatedSrc::new(
             Grid::new(
                 Point2::new(Length::zero(), size_y),
                 Point2::new(1, nr_of_points_y),
@@ -88,7 +88,7 @@ pub fn point_ray_source(cone_angle: Angle, energy: Energy) -> OpmResult<Source> 
         ));
     }
     let size_after_unit_length = (cone_angle / 2.0).tan().value;
-    let light_data_builder = LightDataBuilder::Geometric(RayDataBuilder::PointSrc(PointSrc::new(
+    let light_data_builder = LightDataBuilder::Geometric(RayDataSource::PointSrc(PointSrc::new(
         Hexapolar::new(millimeter!(size_after_unit_length), 3)?.into(),
         UniformDist::new(energy)?.into(),
         LaserLines::new(vec![(nanometer!(1000.0), 1.0)])?.into(),
