+++ opossum_gui/src/components/node_editor/optical_node_editor/properties_editor/light_data_editor/ray_source_editor/distribution_editor/position_distribution_editor/mod.rs
@@ -14,7 +14,7 @@ use grid_editor::GridParam;
 use hexagonal_editor::get_hexagonal_input_params;
 use hexapolar_editor::get_hexapolar_input_params;
 use opossum_core::{
-    position_distributions::PosDistType, prelude::RayDataBuilder,
+    position_distributions::PosDistType, prelude::RayDataSource,
     utils::default_from_name::DefaultFromName,
 };
 use random_editor::RandomParam;
@@ -82,8 +82,8 @@ pub fn NodePosDistInputs(
 #[component]
 pub fn PositionDistributionEditor(
     pos_dist_type: PosDistType,
-    ray_data_builder_sig: ReadSignal<RayDataBuilder>,
-    on_save: EventHandler<RayDataBuilder>,
+    ray_data_builder_sig: ReadSignal<RayDataSource>,
+    on_save: EventHandler<RayDataSource>,
 ) -> Element {
     let mut pos_dist_type_sig = use_signal(|| pos_dist_type);
