+++ opossum_gui/src/components/node_editor/optical_node_editor/properties_editor/light_data_editor/ray_source_editor/ray_type_selection.rs
@@ -4,14 +4,12 @@ use crate::components::node_editor::inputs::{
     input_components::LabeledSelect, select_options_from_enum_iterator,
 };
 use dioxus::prelude::*;
-use opossum_core::{
-    prelude::RayDataBuilder, rays::Rays, utils::default_from_name::DefaultFromName,
-};
+use opossum_core::{prelude::RayDataSource, rays::Rays, utils::default_from_name::DefaultFromName};
 
 #[component]
 pub fn RayDataBuilderSelector(
-    ray_data_builder_sig: ReadSignal<RayDataBuilder>,
-    on_save: EventHandler<RayDataBuilder>,
+    ray_data_builder_sig: ReadSignal<RayDataSource>,
+    on_save: EventHandler<RayDataSource>,
 ) -> Element {
     rsx! {
         LabeledSelect {
@@ -19,11 +17,11 @@ pub fn RayDataBuilderSelector(
             label: "Rays Type",
             options: select_options_from_enum_iterator(
                 &*ray_data_builder_sig.read(),
-                Some(&[&RayDataBuilder::Raw(Rays::default())]),
+                Some(&[&RayDataSource::Raw(Rays::default())]),
             ),
             onchange: move |e: Event<FormData>| {
                 let val = e.value();
-                if let Some(rdb) = RayDataBuilder::default_from_name(val.as_str()) {
+                if let Some(rdb) = RayDataSource::default_from_name(val.as_str()) {
                     on_save.call(rdb);
                 }
             },
